Refactor plot_test_returns into utils.py in HCA
